# swagger.yml
swagger: "2.0"
info:
  title: Entry
  version: 3.0.0
schemes:
  - http
consumes:
  - application/vnd.laincloud.entry.v1+json
produces:
  - application/vnd.laincloud.entry.v1+json
paths:
...
  /api/authorize:
    get:
      tags:
        - auth
      operationId: authorize
      parameters:
        - name: code
          description: "Authorization code"
          in: query
          type: string
          required: true
      responses:
        307:
          description: redirect to homepage of frontend
          headers:
            Location:
              description: homepage of frontend
              type: string
            Set-Cookie:
              description: set access_token in cookie
              type: string
        default:
          description: generic error response
          schema:
            $ref: "#/definitions/error"
...
definitions:
...
  error:
    type: object
    required:
      - message
    properties:
      code:
        type: integer
        format: int64
      message:
        type: string
...